Move updateTimestamps method to UpdateTimestampsTrait
